typecast bugfix

+++ aac/aac.go
@@ -52,7 +52,7 @@ func GetSPF(header []byte) int {
 	return 1024
 }
 
-func SeekTo1StFrame(f os.File) int {
+func SeekTo1StFrame(f os.File) int64 {
 
 	buf := make([]byte, 50000)
 	f.ReadAt(buf, 0)
@@ -97,7 +97,7 @@ func SeekTo1StFrame(f os.File) int {
 			}
 		}
 	}
-	return int(pos)
+	return pos
 }
 
 func GetFrames(f os.File, framesToRead int) ([]byte, error) {
@@ -262,7 +262,7 @@ func GetFileInfo(filename string, br *float64, spf, sr, frames, ch *int) error {
 		return err
 	}
 
-	logger.Log("First frame found at offset: "+strconv.Itoa(firstFramePos), logger.LOG_DEBUG)
+	logger.Log("First frame found at offset: "+strconv.Itoa(int(firstFramePos)), logger.LOG_DEBUG)
 
 	// now having opened the input file, read the fixed header of the
 	// first frame, to get the audio stream's parameters:
@@ -347,7 +347,7 @@ func GetFileInfo(filename string, br *float64, spf, sr, frames, ch *int) error {
 	*frames = frame - 1
 	nsamples := 1024 * *frames
 	playtime := nsamples / *sr
-	*br = float64(fsize-int64(firstFramePos)) / playtime
+	*br = float64((fsize - firstFramePos)) / float64(playtime)
 	*br = *br * 8 / 1000
 
 	logger.Log("frames    : "+strconv.Itoa(*frames), logger.LOG_DEBUG)
